pyproject.toml
[tool.poetry]
name = "Crop Health CrewAI"
version = "0.1.0"
description = "Provide a report on crop health and crop recommendations"
authors = ["Rudraksh Sachin Joshi <rsjoshi0505@gmail.com>", "João Moura <joaomdmoura@mgail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.22.5"}
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"
langchain-groq = "^0.0.1"
bert-extractive-summarizer = "^0.10.1"
torch = "^2.3.0"
jpype1 = "^1.5.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"